modified the examples by replacing complex ones with simple ones
