+++ test/node-interaction.js
@@ -249,14 +249,14 @@ async function launchPage() {
 
 async function validateWithWaitingTime(child, folder, waitingTime, threshold, compare = true) {
   const [browser, page] = await launchPage();
-  const examples = ['pointcloud', 'bezier', 'scatterplot'];
+  const examples = ['pointcloud', 'scatterplot', 'polygon'];
   for (let i = 0; i < 3; i++) {
     console.log(`Begin the ${examples[i]} example`);
     await page.evaluate(() => App.renderToDOM(document.getElementById('app'))); //eslint-disable-line
     await allEvents(page, 1000, threshold, examples[i]); //eslint-disable-line
     await page.evaluate(() => App.nextTestCase()); //eslint-disable-line
     if (compare) {
-      await compareAllImages(examples[i], 0.105);
+      await compareAllImages(examples[i], 0.05);
     } else {
       await createGoldenImage(examples[i]);
     }
